requirements.txt
streamlit
requests
dotenv
down